python/flask/src/requirements.txt
flask
requests
pyjwt
argon2-cffi